Update runtests.jl
The rationale behind the new tests is that,
The first test checks if Zygote+Yao is working properly or not. 
The next one checks if the results are somewhat accurate (from the parameter shift rule and direct Finite Differences)
If those check out, then the loss is tested, which should be below 1. 
I believe that this should work out really well for now (and act as somewhat of a base for future methods added)